isolate merge handlers
