+++ hls4ml/converters/keras_v3/__init__.py
@@ -1,6 +1,7 @@
 from . import conv  # noqa: F401
 from . import core  # noqa: F401
 from . import einsum_dense  # noqa: F401
+from . import merge  # noqa: F401
 from . import pooling  # noqa: F401
 from ._base import registry as layer_handlers
